refactor: separate AddFolderModal
